move things into directories
