+++ .babelrc
@@ -1,0 +1,9 @@
+{
+  'presets': ['env', 'stage-1', 'react'],
+  'plugins': [
+    'transform-decorators-legacy',
+    'transform-class-properties',
+    'transform-strict-mode',
+  ]
+}
+
