+++ package.json
@@ -5,6 +5,9 @@
   "author": "garrettgsb",
   "license": "MIT",
   "dependencies": {
+    "babel-plugin-transform-decorators-legacy": "^1.3.5",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-stage-1": "^6.24.1",
     "react": "^16.3.2",
     "react-dom": "^16.3.2"
   },
@@ -17,7 +20,7 @@
     "webpack-dev-server": "^3.1.4"
   },
   "scripts": {
-    "start": "webpack-dev-server index.js --mode=development",
+    "start": "webpack-dev-server",
     "build": "webpack"
   },
   "babel": {
