+++ src/public/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <title>⾦</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="/styles/style.css">
   </head>
   <body>
     <div id='root'></div>
