+++ src/views/industry.js
