+++ src/views/shop.js
